src/app/service-test/service-test.component.ts
import { Component, OnInit } from '@angular/core';
import { QuoteService } from '../service/test.service';
import { Test } from '../model/service-test';

@Component({
  selector: 'app-service-test',
  templateUrl: './service-test.component.html',
  styleUrls: ['./service-test.component.css']
})
export class ServiceTestComponent implements OnInit {
  public quoteList: Test[];
  public fetchedList: Test[];
  public quoteText: String = null;

  constructor(private service: QuoteService) {}

  ngOnInit() {
    this.quoteList = this.service.getQuote();
    this.service.fetchQuotesFromServer().then((data: Test[]) => {
      this.fetchedList = data;
    });
  }

  createNewQuote() {
    this.service.addNewQuote(this.quoteText);
    this.quoteText = null;
  }

  removeQuote(index) {
    this.service.removeQuote(index);
  }
}
